+++ lib/glimmer-dsl-opal/samples/hello/hello_progress_bar.rb
@@ -91,14 +91,14 @@ class HelloProgressBar
       }
       
       # :vertical SWT style is unsupported for now
-#       progress_bar(:vertical) {
-#         layout_data(:fill, :center, true, false) {
-#           horizontal_span 4
-#         }
-#         minimum <= [@progress_model, :minimum]
-#         maximum <= [@progress_model, :maximum]
-#         selection <= [@progress_model, :selection]
-#       }
+      progress_bar(:vertical) {
+        layout_data(:fill, :center, true, false) {
+          horizontal_span 4
+        }
+        minimum <= [@progress_model, :minimum]
+        maximum <= [@progress_model, :maximum]
+        selection <= [@progress_model, :selection]
+      }
       
       button {
         layout_data(:fill, :center, true, false) {